Support the Go ... recursive path suffix
What
===
Add support for the Go `...` recursive path suffix, so that paths passed
without it are not recursive and paths passed with it are recursive.

Why
===
So that the tool can be used on paths without the tool recursively
looking at all directories under the path.

Note
===
The tool does not support Go import paths, it still only supports file
paths.